Fixed access violation by writing out of bounds on huge string lengths
Increased MYSTRINGS_MAXIMUM_FILE_SIZE by an order of magnitude

+++ DLLSpy/Utils.cpp
@@ -63,7 +63,7 @@ string ExpandPath(const string& sPath)
 
 void TrimString(string& token)
 {
-    for (size_t rIndex = token.length() - 1; rIndex >= 0; rIndex--)
+    for (int rIndex = static_cast<int>(token.length()) - 1; rIndex >= 0; rIndex--)
     {
         for (const char j : sBadChar)
         {
